requirements.txt
matplotlib
pandas
python_telegram_bot
requests
telegram
tqdm
influxdb-client
python-dotenv